Added collided function
This function will handle the movement of a player . Update the mask. Check if the player would collide with something if they were to move in that direction. If they do then you would get the collided object. Regardless of if they do or don't we have to move the player back to the original spot. Then we update the state of the player.